@@ -1,1 +1,1 @@
-ruby-3.2.2
+ruby-3.3.0
