@@ -1,19 +1,19 @@
+# frozen_string_literal: true
+
 class DirectUploadsController < ApplicationController
   skip_before_action :verify_authenticity_token
 
   def create
-    byebug
     blob = ActiveStorage::Blob.create_before_direct_upload!(blob_args)
     render json: direct_upload_json(blob)
   end
 
-  private 
+  private
 
   def direct_upload_json(blob)
     blob.as_json(root: false, methods: :signed_id).merge(service_url: url_for(blob)).merge(direct_upload: {
-      url: blob.service_url_for_direct_upload,
-      headers: blob.service_headers_for_direct_upload
-    })
+                                                                                             url: blob.service_url_for_direct_upload,
+                                                                                             headers: blob.service_headers_for_direct_upload
+                                                                                           })
   end
-    
 end
