@@ -1,13 +1,15 @@
+# frozen_string_literal: true
+
 class SessionsController < ApplicationController
   def create
     user = User.find_by(username: session_params[:username])
-    if user && user.authenticate(session_params[:password])
+    if user&.authenticate(session_params[:password])
       token = issue_token(user)
       render json: { user: UserSerializer.new(user), jwt: token }, status: :accepted
     elsif !user
-      render json: { error: "No such username exists yet." }, status: :not_found
+      render json: { error: 'No such username exists yet.' }, status: :not_found
     else
-      render json: { error: "Invalid username or password" }, status: :unauthorized
+      render json: { error: 'Invalid username or password' }, status: :unauthorized
     end
   end
 
@@ -15,7 +17,7 @@ class SessionsController < ApplicationController
     if logged_in?
       render json: current_user, status: :accepted
     else
-      render json: { error: "User is not logged in/could not be found." }, status: :unauthorized
+      render json: { error: 'User is not logged in/could not be found.' }, status: :unauthorized
     end
   end
 
