@@ -1,4 +1,6 @@
+# frozen_string_literal: true
+
 class Category < ApplicationRecord
-    has_many :cocktails
-    has_many :bartenders, through: :cocktails, class_name: 'User'
+  has_many :cocktails
+  has_many :bartenders, through: :cocktails, class_name: 'User'
 end
