@@ -1,5 +1,7 @@
+# frozen_string_literal: true
+
 class Follow < ApplicationRecord
-    belongs_to :follower, class_name: 'User'
-    belongs_to :followee, class_name: 'User'
-    validates :follower_id, uniqueness: { scope: :followee_id, message: "already exist"}
+  belongs_to :follower, class_name: 'User'
+  belongs_to :followee, class_name: 'User'
+  validates :follower_id, uniqueness: { scope: :followee_id, message: 'already exist' }
 end
