@@ -1,4 +1,5 @@
+# frozen_string_literal: true
+
 class WorkplaceSerializer < ActiveModel::Serializer
   attributes :id, :bar
-  # has_one :bar
 end
