@@ -1,4 +1,6 @@
-require "active_support/core_ext/integer/time"
+# frozen_string_literal: true
+
+require 'active_support/core_ext/integer/time'
 
 Rails.application.configure do
   # Settings specified here will take precedence over those in config/application.rb.
@@ -71,5 +73,5 @@ Rails.application.configure do
 end
 
 Rails.application.routes.default_url_options = {
-    host: "#{ENV['HOST_URL']}"
+  host: (ENV['HOST_URL']).to_s
 }
