@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 class CreateBars < ActiveRecord::Migration[6.1]
   def change
     create_table :bars do |t|
@@ -6,8 +8,8 @@ class CreateBars < ActiveRecord::Migration[6.1]
       t.string :website
       t.float :rating
       t.integer :total_ratings
-      t.text :photos, array:true 
-      t.text :reviews, array:true 
+      t.text :photos, array: true
+      t.text :reviews, array: true
 
       t.timestamps
     end
