@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 # This file is auto-generated from the current state of the database. Instead
 # of editing this file, please use the migrations feature of Active Record to
 # incrementally modify your database, and then regenerate this schema definition.
@@ -10,125 +12,124 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema[7.0].define(version: 2024_03_20_191045) do
+ActiveRecord::Schema[7.0].define(version: 20_240_324_205_805) do
   # These are extensions that must be enabled in order to support this database
-  enable_extension "plpgsql"
+  enable_extension 'plpgsql'
 
-  create_table "active_storage_attachments", force: :cascade do |t|
-    t.string "name", null: false
-    t.string "record_type", null: false
-    t.bigint "record_id", null: false
-    t.bigint "blob_id", null: false
-    t.datetime "created_at", null: false
-    t.index ["blob_id"], name: "index_active_storage_attachments_on_blob_id"
-    t.index ["record_type", "record_id", "name", "blob_id"], name: "index_active_storage_attachments_uniqueness", unique: true
+  create_table 'active_storage_attachments', force: :cascade do |t|
+    t.string 'name', null: false
+    t.string 'record_type', null: false
+    t.bigint 'record_id', null: false
+    t.bigint 'blob_id', null: false
+    t.datetime 'created_at', precision: nil, null: false
+    t.index ['blob_id'], name: 'index_active_storage_attachments_on_blob_id'
+    t.index %w[record_type record_id name blob_id], name: 'index_active_storage_attachments_uniqueness',
+                                                    unique: true
   end
 
-  create_table "active_storage_blobs", force: :cascade do |t|
-    t.string "key", null: false
-    t.string "filename", null: false
-    t.string "content_type"
-    t.text "metadata"
-    t.string "service_name", null: false
-    t.bigint "byte_size", null: false
-    t.string "checksum", null: false
-    t.datetime "created_at", null: false
-    t.index ["key"], name: "index_active_storage_blobs_on_key", unique: true
+  create_table 'active_storage_blobs', force: :cascade do |t|
+    t.string 'key', null: false
+    t.string 'filename', null: false
+    t.string 'content_type'
+    t.text 'metadata'
+    t.string 'service_name', null: false
+    t.bigint 'byte_size', null: false
+    t.string 'checksum', null: false
+    t.datetime 'created_at', precision: nil, null: false
+    t.index ['key'], name: 'index_active_storage_blobs_on_key', unique: true
   end
 
-  create_table "active_storage_variant_records", force: :cascade do |t|
-    t.bigint "blob_id", null: false
-    t.string "variation_digest", null: false
-    t.index ["blob_id", "variation_digest"], name: "index_active_storage_variant_records_uniqueness", unique: true
+  create_table 'active_storage_variant_records', force: :cascade do |t|
+    t.bigint 'blob_id', null: false
+    t.string 'variation_digest', null: false
+    t.index %w[blob_id variation_digest], name: 'index_active_storage_variant_records_uniqueness', unique: true
   end
 
-  create_table "bars", force: :cascade do |t|
-    t.string "name"
-    t.string "address"
-    t.string "website"
-    t.float "rating"
-    t.integer "total_ratings"
-    t.text "photos", array: true
-    t.text "reviews", array: true
-    t.datetime "created_at", null: false
-    t.datetime "updated_at", null: false
+  create_table 'bars', force: :cascade do |t|
+    t.string 'name'
+    t.string 'address'
+    t.string 'website'
+    t.float 'rating'
+    t.integer 'total_ratings'
+    t.text 'photos', array: true
+    t.text 'reviews', array: true
+    t.datetime 'created_at', null: false
+    t.datetime 'updated_at', null: false
   end
 
-  create_table "categories", force: :cascade do |t|
-    t.string "name"
-    t.string "definition"
-    t.jsonb "popular_drinks", default: [], array: true
-    t.datetime "created_at", null: false
-    t.datetime "updated_at", null: false
+  create_table 'categories', force: :cascade do |t|
+    t.string 'name'
+    t.string 'definition'
+    t.jsonb 'popular_drinks', default: [], array: true
+    t.datetime 'created_at', null: false
+    t.datetime 'updated_at', null: false
   end
 
-  create_table "cocktails", force: :cascade do |t|
-    t.string "name"
-    t.string "image"
-    t.string "description"
-    t.string "execution"
-    t.text "ingredients", array: true
-    t.bigint "bartender_id"
-    t.bigint "category_id", null: false
-    t.datetime "created_at", null: false
-    t.datetime "updated_at", null: false
-    t.integer "likes_count", default: 0, null: false
-    t.index ["bartender_id"], name: "index_cocktails_on_bartender_id"
-    t.index ["category_id"], name: "index_cocktails_on_category_id"
+  create_table 'cocktails', force: :cascade do |t|
+    t.string 'name'
+    t.string 'description'
+    t.string 'execution'
+    t.text 'ingredients', array: true
+    t.bigint 'bartender_id'
+    t.bigint 'category_id', null: false
+    t.datetime 'created_at', null: false
+    t.datetime 'updated_at', null: false
+    t.integer 'likes_count', default: 0, null: false
+    t.index ['bartender_id'], name: 'index_cocktails_on_bartender_id'
+    t.index ['category_id'], name: 'index_cocktails_on_category_id'
   end
 
-  create_table "favorites", force: :cascade do |t|
-    t.bigint "user_id", null: false
-    t.integer "favorited_bar_id"
-    t.datetime "created_at", null: false
-    t.datetime "updated_at", null: false
-    t.index ["user_id"], name: "index_favorites_on_user_id"
+  create_table 'favorites', force: :cascade do |t|
+    t.bigint 'user_id', null: false
+    t.integer 'favorited_bar_id'
+    t.datetime 'created_at', null: false
+    t.datetime 'updated_at', null: false
+    t.index ['user_id'], name: 'index_favorites_on_user_id'
   end
 
-  create_table "follows", force: :cascade do |t|
-    t.integer "follower_id"
-    t.integer "followee_id"
-    t.datetime "created_at", null: false
-    t.datetime "updated_at", null: false
+  create_table 'follows', force: :cascade do |t|
+    t.integer 'follower_id'
+    t.integer 'followee_id'
+    t.datetime 'created_at', null: false
+    t.datetime 'updated_at', null: false
   end
 
-  create_table "likes", force: :cascade do |t|
-    t.integer "liker_id"
-    t.integer "liked_cocktail_id"
-    t.datetime "created_at", null: false
-    t.datetime "updated_at", null: false
+  create_table 'likes', force: :cascade do |t|
+    t.integer 'liker_id'
+    t.integer 'liked_cocktail_id'
+    t.datetime 'created_at', null: false
+    t.datetime 'updated_at', null: false
   end
 
-  create_table "users", force: :cascade do |t|
-    t.string "full_name"
-    t.string "username"
-    t.string "password_digest"
-    t.string "location"
-    t.boolean "bartender"
-    t.string "instagram_account"
-    t.string "biography"
-    t.integer "insta_follower"
-    t.integer "insta_following"
-    t.text "profile_pic"
-    t.datetime "created_at", null: false
-    t.datetime "updated_at", null: false
+  create_table 'users', force: :cascade do |t|
+    t.string 'full_name'
+    t.string 'username'
+    t.string 'password_digest'
+    t.string 'location'
+    t.boolean 'bartender'
+    t.string 'instagram_account'
+    t.string 'biography'
+    t.integer 'insta_follower'
+    t.integer 'insta_following'
+    t.datetime 'created_at', null: false
+    t.datetime 'updated_at', null: false
   end
 
-  create_table "workplaces", force: :cascade do |t|
-    t.bigint "bar_id", null: false
-    t.bigint "bartender_id"
-    t.string "schedule"
-    t.datetime "created_at", null: false
-    t.datetime "updated_at", null: false
-    t.index ["bar_id"], name: "index_workplaces_on_bar_id"
-    t.index ["bartender_id"], name: "index_workplaces_on_bartender_id"
+  create_table 'workplaces', force: :cascade do |t|
+    t.bigint 'bar_id', null: false
+    t.bigint 'bartender_id'
+    t.string 'schedule'
+    t.datetime 'created_at', null: false
+    t.datetime 'updated_at', null: false
+    t.index ['bar_id'], name: 'index_workplaces_on_bar_id'
+    t.index ['bartender_id'], name: 'index_workplaces_on_bartender_id'
   end
 
-  add_foreign_key "active_storage_attachments", "active_storage_blobs", column: "blob_id"
-  add_foreign_key "active_storage_variant_records", "active_storage_blobs", column: "blob_id"
-  add_foreign_key "cocktails", "categories"
-  add_foreign_key "cocktails", "users", column: "bartender_id"
-  add_foreign_key "favorites", "users"
-  add_foreign_key "workplaces", "bars"
-  add_foreign_key "workplaces", "users", column: "bartender_id"
+  add_foreign_key 'active_storage_attachments', 'active_storage_blobs', column: 'blob_id'
+  add_foreign_key 'active_storage_variant_records', 'active_storage_blobs', column: 'blob_id'
+  add_foreign_key 'cocktails', 'categories'
+  add_foreign_key 'cocktails', 'users', column: 'bartender_id'
+  add_foreign_key 'favorites', 'users'
+  add_foreign_key 'workplaces', 'bars'
+  add_foreign_key 'workplaces', 'users', column: 'bartender_id'
 end
